Various changes to spectrum & spectrumBuilder.
[Details]
- Added a default duplicate strategy for the SpectrumBuilder
- Changed Spectrum Element's to now fully rely on the Detail object to
  store the details of that element. This reduces space requirements and
  makes the implementation much more readable.
- Moved the spectrumBuilder module to the flitsr.input package. A dummy
  file has been kept in place of the old one so that commands like
  `from flitsr.spectrumBuilder import SpectrumBuilder` still work.
- Exposed the Input, TCM and Gzoltar input objects directly from
  flitsr.input now, so that for example `from flitsr.input import Input`
  and `from flitsr.input import TCM` can be used (the old format can still
  be used as well).

[Note]
- Updated the docs with the changes to the API.

[File's changed]
	modified:   docs/source/api.rst
	modified:   docs/source/input_types.rst
	modified:   flitsr/__init__.py
	modified:   flitsr/input/__init__.py
	modified:   flitsr/input/gzoltar_input.py
	modified:   flitsr/input/input_reader.py
	new file:   flitsr/input/spectrumBuilder.py
	modified:   flitsr/input/tcm_input.py
	modified:   flitsr/spectrum.py
	modified:   flitsr/spectrumBuilder.py
