+++ flitsr/__init__.py
@@ -1,3 +1,5 @@
+from flitsr.input import spectrumBuilder
+
 __all__ = [
     'flitsr',
     'merge',
